refactor clock
