+++ src/node/clock/tick/watch.go
@@ -1,4 +1,4 @@
-package node
+package tick
 
 import "time"
