+++ test/node/clock/tick/engine_test.go
@@ -1,10 +1,10 @@
-package protocol
+package tick
 
 import (
 	"fmt"
-	"github.com/my-cloud/ruthenium/src/clock/node"
+	"github.com/my-cloud/ruthenium/src/node/clock/tick"
 	"github.com/my-cloud/ruthenium/test"
-	"github.com/my-cloud/ruthenium/test/clock/clocktest"
+	"github.com/my-cloud/ruthenium/test/node/clock/clocktest"
 	"testing"
 	"time"
 )
@@ -14,7 +14,7 @@ func Test_Do(t *testing.T) {
 	watchMock := new(clocktest.WatchMock)
 	watchMock.NowFunc = func() time.Time { return time.Unix(0, 0) }
 	var calls int
-	engine := node.NewEngine(func(int64) { calls++ }, watchMock, 1, 0, 0, nil)
+	engine := tick.NewEngine(func(int64) { calls++ }, watchMock, 1, 0, 0, nil)
 
 	// Act
 	engine.Do()
