+++ test/node/network/p2p/host_test.go
@@ -4,8 +4,8 @@ import (
 	gp2p "github.com/leprosus/golang-p2p"
 	"github.com/my-cloud/ruthenium/src/node/network/p2p"
 	"github.com/my-cloud/ruthenium/test"
-	"github.com/my-cloud/ruthenium/test/clock/clocktest"
 	"github.com/my-cloud/ruthenium/test/log/logtest"
+	"github.com/my-cloud/ruthenium/test/node/clock/clocktest"
 	"github.com/my-cloud/ruthenium/test/node/network/networktest"
 	"github.com/my-cloud/ruthenium/test/node/network/p2p/p2ptest"
 	"github.com/my-cloud/ruthenium/test/node/protocol/protocoltest"
