+++ test/node/protocol/validation/transactions_pool_test.go
@@ -8,8 +8,8 @@ import (
 	"github.com/my-cloud/ruthenium/src/node/protocol/validation"
 	"github.com/my-cloud/ruthenium/src/ui/server"
 	"github.com/my-cloud/ruthenium/test"
-	"github.com/my-cloud/ruthenium/test/clock/clocktest"
 	"github.com/my-cloud/ruthenium/test/log/logtest"
+	"github.com/my-cloud/ruthenium/test/node/clock/clocktest"
 	"github.com/my-cloud/ruthenium/test/node/protocol/protocoltest"
 	"testing"
 	"time"
